feat: Get the nex sequence number in holiday

+++ src/main/java/com/iws_manager/iws_manager_api/repositories/PublicHolidayRepository.java
@@ -14,5 +14,5 @@ public interface PublicHolidayRepository extends JpaRepository<PublicHoliday, Lo
     List<PublicHoliday> findAllByOrderBySequenceNoDesc();
 
     @Query("SELECT COALESCE(MAX(p.sequenceNo), 0) FROM PublicHoliday p")
-    Integer findMaxSequenceNo();
+    Long findMaxSequenceNo();
 }
